+++ src/GenomeTrackSummary.cpp
@@ -284,14 +284,14 @@ SEXP gintervals_summary(SEXP _expr, SEXP _intervals, SEXP _iterator_policy, SEXP
 
 			if (scanner.get_iterator()->is_1d()) {
 				GIntervalsBigSet1D::begin_save(intervset_out.c_str(), iu, chromstats1d);
-				for (int chromid = 0; chromid < iu.get_chromkey().get_num_chroms(); ++chromid) {
+				for (int chromid = 0; (uint64_t)chromid < iu.get_chromkey().get_num_chroms(); ++chromid) {
 					if (intervals1d->size(chromid))
 						chroms1d.insert(chromid);
 				}
 			} else {
 				GIntervalsBigSet2D::begin_save(intervset_out.c_str(), iu, chromstats2d);
-				for (int chromid1 = 0; chromid1 < iu.get_chromkey().get_num_chroms(); ++chromid1) {
-					for (int chromid2 = 0; chromid2 < iu.get_chromkey().get_num_chroms(); ++chromid2) {
+				for (int chromid1 = 0; (uint64_t)chromid1 < iu.get_chromkey().get_num_chroms(); ++chromid1) {
+					for (int chromid2 = 0; (uint64_t)chromid2 < iu.get_chromkey().get_num_chroms(); ++chromid2) {
 						if (intervals2d->size(chromid1, chromid2))
 							chroms2d.insert(ChromPair(chromid1, chromid2));
 					}
